+++ Kipu.Platform/Alerts/Interfaces/Rest/Resources/CreateAlertResource.cs
@@ -7,7 +7,7 @@ public record CreateAlertResource(
     string Type,
     string Severity,
     string Message,
-    int CurrentStock,
-    int MinStock,
+    decimal CurrentStock,
+    decimal MinStock,
     int? DaysToExpiry,
     int? WarehouseId = null);
