+++ Kipu.Platform/Dashboard/Interfaces/Rest/Resources/TopStockProductResource.cs
@@ -1,3 +1,3 @@
 namespace Kipu.Platform.Dashboard.Interfaces.Rest.Resources;
 
-public record TopStockProductResource(int ProductId, string ProductName, int TotalStock);
+public record TopStockProductResource(int ProductId, string ProductName, decimal TotalStock);
