+++ Kipu.Platform/Products/Domain/Model/Commands/RegisterStockSaleCommand.cs
@@ -6,4 +6,4 @@ namespace Kipu.Platform.Products.Domain.Model.Commands;
 ///     once Sales exists, by an event handler reacting to a domain event it
 ///     raises), never called directly by the frontend.
 /// </summary>
-public record RegisterStockSaleCommand(int ProductId, int BusinessId, int Quantity);
+public record RegisterStockSaleCommand(int ProductId, int BusinessId, decimal Quantity);
