+++ Kipu.Platform/Products/Interfaces/Rest/Resources/RegisterStockIntakeResource.cs
@@ -3,10 +3,10 @@ namespace Kipu.Platform.Products.Interfaces.Rest.Resources;
 /// <summary>ProductId is deliberately not part of this resource — it comes from the route (/products/{id}/stock-intake).</summary>
 public record RegisterStockIntakeResource(
     int WarehouseId,
-    int Quantity,
+    decimal Quantity,
     decimal? PurchasePrice,
     DateOnly? Expiration,
     string? Supplier,
     string? Note,
-    int? MinimumStock,
+    decimal? MinimumStock,
     int? SupplierId = null);
